Ctrl and Alt keys not treated correctly in Input
event.ctrlKey and event.altKey are booleans that represent if ctrl and alt are pressed while the current key that triggered the event is pressed (or released). It does not represent the fact that the ctrl or alt key is the key that is pressed. To do so, we have to use the corresponding keycode.

In the previous code, self.ctrlDown was never set to false because when ctrl is up, event.ctrlKey is false although the key code is 17 (the CTRL key code). Indeed, the ctrl key is not pressed at this moment, because it is released. (idem for the alt key)

Also, the ctrl and alt key are now added to the this.keyDown array, like any other keys. The previous if / else statement prevents that.