Update KeyboardPanRotateDollyHandler.js
Some user input filtering logic is already made by scene input (input.js). Camera handlers should depend on scene input and not document events.

Example of some issue: if the user focus is on an INPUT or TEXTAREA element, it must be impossible to pan the model even if the mouse is hovering it. In input.js, this logic is correctly handled and the event is not fired. However, as this handler listens on the document events and not scene input, the model moves.