[data-textures]: implement a smart deferred flags update mechanism
If the number of updated per-object properties in a frame is less than a threshold, they will be updated directly.

If the number of updates is instead bigger than a threshold (by default 10), the ones above the threshold will be batched, and the batch flushed just before rendering the frame.

This completely solves https://github.com/xeokit/xeokit-sdk/issues/1033